requirements.txt
streamlit
tensorflow
fpdf
Pillow
numpy
